Improve start-up time by defering slow imports.
The `configurator` package takes 200-300 ms to import, which is enough to
cause a noticeable pause when a program is starting up.  Almost all of
this time is spent importing `pkg_resources`:

$ python -X importtime -c 'import configurator'
$ python -X importtime -c 'import pkg_resources'

By defering this import until it is needed, `configurator` can start up
in about 7 ms---much faster.  This is really nice if the command being
executed doesn't actually need to read a config file, e.g. if it's just
displaying usage help or something.